+++ src/routes/projects.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import User from '../controllers/projects';
+
+const router = express.Router();
+
+router.route('/user/:username').get(User.ownedProjects);
+
+module.exports = router;
