--- src/routes/user.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import User from '../controllers/user';
-
-const router = express.Router();
-
-router.route('/projects/:username').get(User.ownedProjects);
-
-module.exports = router;
